backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
torch>=2.0.0
torchvision
transformers>=4.30.0
Pillow>=9.0.0
sentencepiece>=0.1.99
accelerate
python-multipart
